Extracted the send-message actions to a function
A couple commits prior to this, there was a separate procedure
for sending a message, but its boundary was on the other side
of the for loop, requiring a generator/iterator, which was
non-necessary complexity.
After merging that procedure into its caller, I saw a better boundary.
Now, I made a new send-message procedure that performs one action
and includes the for-loop.
Also added docstrings to the procedures.