Make llmbot event-driven
The issue was that sending >1 LLM replies
resulted in only the last one reaching its destination.
I estimate the root cause to be that sending messages
from within the on_receive() callback causes problems within
the meshtastic client.

This fix is to transition the llmbot app to be an event-driven
state machine.  It receives and enqueues the message, then
sends reply fragment(s) in a context outside of on_receive().
This requires keeping track of both the message and sender;
especially because message fragments are sent with a long delay
between them during which other messages may arrive.