Reply with the packet ID
When a radio pkt is received, the script captures the
packet ID (an arbitrary int) and associates it with the message.
The packet ID is used when sending the LLM reply out
on the radio.  This way, the recipient's client can show
which user message the reply is in reference to.

This capability is not essential to function, but it can allow a client
software to help organize or display messages in a more organized
fashion.