Fix fragmented messages
The problem was that users were only getting the final
fragment of a multi-part message.
The fix was to add time between sending the messages.
20 seconds was not enough.
So the delay between parts is 30 seconds.

I also removed the unicode ellipses that was indicating
successive message fragment since it wasn't displayed
properly on the client.